Add unit movement

+++ src/awrs/cell.rs
@@ -1,6 +1,5 @@
-
 #[derive(Clone)]
 pub struct Cell {
-    pub x: u32,
-    pub y: u32,
+    pub x: usize,
+    pub y: usize,
 }
